requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1

SQLAlchemy==2.0.36
pymysql==1.1.1

APScheduler==3.10.4
python-dateutil==2.9.0.post0

Jinja2==3.1.4
markdown==3.7

pandas==2.2.2

email-validator==2.2.0